Week one of seminar on landscape evolution modeling.
Beginnings of a landscape evolution model designed to be equivalent to
the Fastscape model of Braun and Willet (2013) (doi:
10.1016/j.geomorph.2012.10.008). This iteration generates the initial
nx by ny landscape. It assigns elevations of zero to the edges of the
model domain and positive, random numbers as elevations of all other
nodes. Then, we assign the receiver array “rec,” which identifies the
node which will receive water from the current node (defined by the
path of steepest descent). We also calculate the distance between each
node and its receiver node in the array “length.”